\name{author}
\docType{data}
\alias{author}
\title{Author dataset}
\description{This data matrix contains the counts of the 26 letters of the alphabet (columns of matrix) for 12 different novels (rows of matrix). Each row contains letter counts in a sample of text from each work, excluding proper nouns.}
\usage{data("author")}
\format{Data frame containing the 12 x 26 matrix.}
\source{Larsen, W.A. and McGill, R., unpublished data collected in 1973.}
\keyword{datasets}
