\name{wg93}
\docType{data}
\alias{wg93}
\title{International Social Survey Program on Environment 1993 - western German sample}
\description{This data frame contains records of four questions on attitude towards science with responses on a five-point scale (1=agree strongly to 5=disagree strongly) and three demographic variables (sex, age and education).}
\usage{data(wg93)}
\format{Data frame (871x7).}
\source{ISSP (1993). International Social Survey Program: Environment. \url{http://www.issp.org}}
\keyword{datasets}
