\name{pchlist}
\alias{pchlist}
\title{Listing the set of available symbols.}
\description{A plot of the available symbols for use with the option \code{pch}.}
\usage{pchlist() }
\details{
This function generates a numbered list of the plotting symbols available for use in the functions \code{\link{plot.ca}} and \code{\link{plot3d.ca}}.
}
\seealso{\code{\link{plot.ca}}, \code{\link{plot3d.ca}}}
\examples{
pchlist()
}

