\name{print.summary.ca}
\alias{print.summary.ca}
\title{Printing summeries of ca objects}
\description{Printing method for summaries of correspondence analysis objects}
\usage{\method{print}{summary.ca}(x, ...) }
\arguments{
  \item{x}{Summary of a simple correspondence analysis object returned by \code{\link{summary.ca}}}
  \item{...}{Further arguments are ignored}
          }
\seealso{\code{\link{ca}}, \code{\link{summary.ca}}}
