\name{print.summary.mjca}
\alias{print.summary.mjca}
\title{Printing summeries of mjca objects}
\description{Printing method for summeries of multiple and joint correspondence analysis objects}
\usage{\method{print}{summary.mjca}(x, ...) }
\arguments{
  \item{x}{summary of a multiple or joint correspondence analysis object returned by \code{\link{summary.mjca}}}
  \item{...}{Further arguments are ignored}
          }
\details{}

\references{}
\seealso{\code{\link{mjca}}, \code{\link{summary.mjca}}}
\examples{}
\keyword{}
