\name{smoke}
\docType{data}
\alias{smoke}
\title{Smoke dataset}
\description{Artificial dataset in Greenacre (1984)}
\usage{data(smoke)}
\format{Table containing 5 rows (staff group) and 4 columns (smoking categories), giving the frequencies of smoking categories in each staff group in a fictional organization.}
\references{Greenacre, M.J. (1984). \emph{Theory and Applications of Correspondence Analysis}. London: Academic Press.}
\keyword{datasets}
